Move CacheImageInTheBackground to registry
